# Rewrite _IGzipReader in C for even less overhead.