--- a/setup.py
+++ b/setup.py
@@ -158,7 +158,6 @@ setup(
     classifiers=[
         "Programming Language :: Python :: 3 :: Only",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
@@ -173,6 +172,6 @@ setup(
         "Operating System :: MacOS",
         "Operating System :: Microsoft :: Windows",
     ],
-    python_requires=">=3.7",  # We use METH_FASTCALL
+    python_requires=">=3.8",  # BadGzipFile imported
     ext_modules=EXTENSIONS
 )