--- a/src/isal/igzip.py
+++ b/src/isal/igzip.py
@@ -168,7 +168,7 @@ class IGzipFile(gzip.GzipFile):
                                                   0)
         if self.mode == READ:
             raw = _IGzipReader(self.fileobj)
-            self._buffer = io.BufferedReader(raw, 128 * 1024)
+            self._buffer = io.BufferedReader(raw, 64 * 1024)
 
     def __repr__(self):
         s = repr(self.fileobj)