… .github/workflows/ci.yml
… tests/test_gzip_compliance.py
… tests/test_igzip.py