### Checklist
- [ ] Pull request details were added to CHANGELOG.rst
- [ ] Documentation was updated (if needed)

This was primarily motivated by trying to write an implementation of the reader that escapes the GIL and reduces wall clock time. It worked, except that the wall clock time was not much reduced. I decided to fully implement the _GzipReader in C to reduce python overhead. It is quite successful

Before:
```
Benchmark 1: python -m isal.igzip -cd ~/test/5millionreads_R1.fastq.gz > /dev/null
  Time (mean ± σ):      2.151 s ±  0.005 s    [User: 2.106 s, System: 0.045 s]
  Range (min … max):    2.138 s …  2.158 s    10 

Benchmark 1: python -m isal.igzip -cd ~/test/nanopore_100000reads.u.bam > /dev/null
  Time (mean ± σ):     442.3 ms ±  12.3 ms    [User: 349.2 ms, System: 93.1 ms]
  Range (min … max):   432.3 ms … 468.9 ms    10 runs
 
Benchmark 1: python -m isal.igzip -cd ~/test/nanopore_100000reads.5.bam > /dev/null
  Time (mean ± σ):      2.916 s ±  0.015 s    [User: 2.844 s, System: 0.072 s]
  Range (min … max):    2.881 s …  2.937 s    10 runs
```

After:
```
Benchmark 1: python -m isal.igzip -cd ~/test/5millionreads_R1.fastq.gz > /dev/null
  Time (mean ± σ):      1.980 s ±  0.008 s    [User: 1.936 s, System: 0.044 s]
  Range (min … max):    1.966 s …  1.995 s    10 runs
 
Benchmark 1: python -m isal.igzip -cd ~/test/nanopore_100000reads.u.bam > /dev/null
  Time (mean ± σ):     261.4 ms ±   1.9 ms    [User: 159.7 ms, System: 101.6 ms]
  Range (min … max):   259.2 ms … 264.3 ms    10 runs

Benchmark 1: python -m isal.igzip -cd ~/test/nanopore_100000reads.5.bam > /dev/null
  Time (mean ± σ):      2.748 s ±  0.014 s    [User: 2.671 s, System: 0.076 s]
  Range (min … max):    2.722 s …  2.772 s    10 runs
```
For comparison, the igzip run.
```
Benchmark 1: igzip -cd ~/test/5millionreads_R1.fastq.gz > /dev/null
  Time (mean ± σ):      1.946 s ±  0.003 s    [User: 1.910 s, System: 0.036 s]
  Range (min … max):    1.939 s …  1.950 s    10 runs

Benchmark 1: igzip -cd ~/test/nanopore_100000reads.u.bam > /dev/null
  Time (mean ± σ):     238.9 ms ±   2.4 ms    [User: 128.5 ms, System: 110.3 ms]
  Range (min … max):   236.0 ms … 242.2 ms    10 runs
 
Benchmark 1: igzip -cd ~/test/nanopore_100000reads.5.bam > /dev/null
  Time (mean ± σ):      2.603 s ±  0.014 s    [User: 2.524 s, System: 0.078 s]
  Range (min … max):    2.580 s …  2.629 s    10 run
```

Execution time vs pure C
before: 2151/1946 = 110.5 % -> 10.5% overhead 
Overhead after 2024/1940 = 101.7% -> 1.7 % overhead. 
Quite great.

For BGZIP compression as used in BAM it also works wonders.

EDIT: Updated with latest benchmark numbers.